Crate cplwm_assignment

Root file of your solution

This file

There are two types of projects or crates in Rust: libraries and executables. The difference should be obvious. The root of a library project is src/lib.rs (like this file) and the root of an executable project is src/main.rs.

The root file contains some important things:

It provides some documentation for the crate. You are reading it right now.

It contains annotations that are true for the whole crate. See the source code of this file for an example.

It declares dependencies on external crates. Again, see the source code.

It declares the modules of which the crate consists using the mod keyword. Don't confuse this with the use keyword. If you add any modules to this project, you have to declare them at the end of this file using the mod keyword. Unless they are submodules of another module, in which case you have to declare them at the top of that module. In case you want them to be public (and visible in the documentation too), use pub mod instead of mod.

See the chapter on Crates and Modules in the Rust book for more information.

The source code of this file contains more information.

If you are looking at the documentation of this file, the list on the left side contains all the crates that this crate transitively depends on.

The next thing to look at is the a_fullscreen_wm module.

Rust Version

TODO: indicate which version of Rust you are using. If you are using 1.12.1, you don't have to do anything. Otherwise, replace the version below with the output of rustc --version.

VERSION: rustc 1.12.1 (d4f39402a 2016-10-19)

If you are using a nightly as opposed to a stable version of Rust, explain why below:

...

Modules

a_fullscreen_wm
Fullscreen Window Manager

b_tiling_wm
Tiling Window Manager

c_floating_windows
Floating Windows

d_minimising_windows
Minimising Windows

e_fullscreen_windows
Optional: Fullscreen Windows

f_gaps
Optional: Gaps

g_multiple_workspaces
Optional: Multiple Workspaces

h_different_tiling_layout
Optional: Different Tiling Layout