Enum cplwm_assignment::a_fullscreen_wm::FullscreenWMError

pub enum FullscreenWMError {
    UnknownWindow(Window),
    ManagedWindow(Window),
}

The errors that this window manager can return.

For more information about why you need this, read the documentation of the associated Error type of the WindowManager trait.

In the code below, we would like to return an error when we are asked to do something with a window that we do not manage, so we define an enum FullscreenWMError with one variant: UnknownWindow.

Feel free to add or remove variants from this enum. You may also replace it with a type or struct if you wish to do so.

Variants
UnknownWindow(Window)

This window is not known by the window manager.

ManagedWindow(Window)

This window is managed by the windows manager right now which can't be added again.