Struct cplwm_assignment::a_fullscreen_wm::FullscreenWM

pub struct FullscreenWM {
    pub windows: Vec<Window>,
    pub screen: Screen,
    pub windows_type: Vec<(Window, FloatOrTile)>,
    pub focused_window: Option<Window>,
}

The FullscreenWM struct

The first thing to do when writing a window manager, is to define a struct (or enum) that will contain the state of the window manager, e.g. the managed windows along with their geometries, the focused window, etc.

Depending on the layout and the functionality the window manager provides, this can vary from simple Vecs to trees, hashmaps, etc. You can have a look at the collections Rust provides.

Remember that you are free to add additional dependencies to your project, e.g., for another type of data structure. But this is certainly not required. For more information, see the Hints & Tricks section of the assignment.

Example Representation

The fullscreen window manager that we are implementing is very simple: it just needs to keep track of all the windows that were added and remember which one is focused. It is not even necessary to remember the geometries of the windows, as they will all be resized to the size of the screen.

A possible data structure to keep track of the windows is a simple Vec: the last element in the vector is the window on top, which is also the only window to display. Why not the first element? Because it is easier to add an element to the end of a vector. This is convenient, as adding a new window should also put it on top of the other windows.

Another thing we need to keep track of is the Screen, because we must resize the windows the size of the screen. A Screen is passed via the new method of the trait and the resize_screen method of the trait updates the screen with a new one.

These two fields are enough to get started, which does not mean that they are enough to correctly implement this window manager. As you will notice in a short while, there is a problem with this representation. Feel free to add/replace/remove fields.

To understand the #derive[(..)] line before the struct, read the Supertraits section of the WindowManager trait.

Fields
windows: Vec<Window>

A vector of windows, the first one is on the bottom, the last one is on top, and also the only visible window.

screen: Screen

We need to know which size the fullscreen window must be.

windows_type: Vec<(Window, FloatOrTile)>

keep track of the type after add function

focused_window: Option<Window>

In oreder to meet the requirment of , when there is no window focused, but the windows are still under control. add a new variable to keep track of focused window