Module cplwm_assignment::b_tiling_wm

Tiling Window Manager

Write a more complex window manager that will tile its windows. Tiling is described in the first section of the assignment. Your window manager must implement both the WindowManager trait and the TilingSupport trait. See the documentation of the TilingSupport trait for the precise requirements and an explanation of the tiling layout algorithm.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: YES

COMMENTS: I don't save the type of the window here,because all are tiles when call the fucntion get_window_info

Structs

TilescreenWM
Stuct of Tile screen windows manager

Enums

TilescreenWMError
The error for tile screen

Type Definitions

WMName
TilescreenWM