Struct cplwm_assignment::b_tiling_wm::TillingWM

pub struct TillingWM {
    pub windows: VecDeque<WindowWithInfo>,
    pub screen: Screen,
}

The TillingWM struct

Example Representation

Similar as previous exercises, once again I use VedDeue due to the power to push and pop in both sides of the vector, likewise, it stores the whole window_with_info which will keep track the invidiaul geometry of each window, screen was kept as previous exercise

Fields
windows: VecDeque<WindowWithInfo>

TODO: Documentation

screen: Screen

TODO: Documentation