Struct cplwm_assignment::c_floating_windows::FloatingWM

pub struct FloatingWM {
    pub windows: Vec<WindowWithInfo>,
    pub screen: Screen,
    pub index_foused_window: Option<usize>,
}

The FloatingWM struct

Example Representation

Similar as previous exercises

Fields
windows: Vec<WindowWithInfo>

TODO: No need to used a complex collection

screen: Screen

TODO: Documentation

index_foused_window: Option<usize>

The index of the focused window in the collection, if there is no focused window a None is placed

Methods

impl FloatingWM

Supported functions

fn get_next_tile_index(&self, index: usize, saved: usize) -> Option<usize>

Get the next index by the given index of the tiled windows.

fn get_prev_tile_index(&self, index: usize, saved: usize) -> Option<usize>

Get the previous index by the given index of the tiled windows.