Module cplwm_assignment::d_minimising_windows

Minimising Windows

Extend your window manager with support for (un)minimising windows. i.e. the ability to temporarily hide windows and to reveal them again later. See the documentation of the MinimiseSupport trait for the precise requirements.

Either make a copy of the tiling window manager with support for floating windows you developed in the previous assignment and let it implement the MinimiseSupport trait as well, or implement this trait by building a wrapper around the previous window manager. Note that this window manager must still implement all the traits from previous assignments.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: PARTIAL

COMMENTS:

it is a requesto to keep track of the position of them minised elements, the best way I concude to actually do this ordering is with a collection, in this case a vec is use to keep track the position minised elements, such vector it is par of the MinimisingWM structure. Moreover it is important to remove the minised element from the windows vector, however if the removed is used as it was implemented in the others window managers, we actually deleted from the windows vector which will be actually

collection of minimised windows, this is one approach in which I can keep the order as I did with the floating and tiled windows so the approach is that MinimisingWM will have another attribute called minimised_windows which will be a vec of windows this vec will keep the order in which the windows are minimised. Once again get_window_layout should be updated to filter out the minised windows in order to show only the unminised windows.

Structs

MinimisingWM
The MinimisingWM struct

Enums

MinimisingWMError
TODO: Documentation

Type Definitions

WMName
TODO: Documentation