Module cplwm_assignment::e_fullscreen_windows

Optional: Fullscreen Windows

Extend your window manager with support for fullscreen windows, i.e. the ability to temporarily make a window take up the whole screen, thereby obscuring all other windows. See the documentation of the FullscreenSupport trait for the precise requirements. Don't confuse this with the first assignment, in which you built a window manager that displayed all windows fullscreen.

Like in the previous assignments, either make a copy of, or define a wrapper around your previous window manager to implement the FullscreenSupport trait as well. Note that this window manager must still implement all the traits from previous assignments.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: YES

COMMENTS:

Structs

FloatscreenWM
Stuct of Tile screen windows manager

Enums

FloatscreenWMError
The error for tile screen

Type Definitions

WMName
FloatscreenWM