Module cplwm_assignment::f_gaps

Optional: Gaps

Extend your window manager with support for gaps, i.e. the ability to add some space between the different tiles. See the documentation of the GapSupport trait for the precise requirements.

Make a copy of your tiling window manager from assignment B and let it implement the GapSupport trait. You are not required to let this window manager implement all the previous traits.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: YES

COMMENTS:

...

Structs

FloatscreenWM
Stuct of Tile screen windows manager

Enums

FloatscreenWMError
The error for tile screen

Type Definitions

WMName
FloatscreenWM