Struct cplwm_assignment::f_gaps::FloatscreenWM

pub struct FloatscreenWM {
    pub windows: Vec<Window>,
    pub full_screen_windows: Vec<(Window, Geometry, FloatOrTile, Option<Window>)>,
    pub focused_window: Option<Window>,
    pub float_windows: Vec<(Window, Geometry)>,
    pub tile_windows: Vec<Window>,
    pub screen: Screen,
    pub original_windows: Vec<(Window, Geometry)>,
    pub minimised_windows: Vec<(Window, Geometry, FloatOrTile)>,
    pub gap: GapSize,
}

Stuct of Tile screen windows manager

Fields
windows: Vec<Window>

The vectors of the windows, the first one is on the master tile

full_screen_windows: Vec<(Window, Geometry, FloatOrTile, Option<Window>)>

The full screen windows,contains all full_screen windows, but only the last one is visible. when you close all full screen windows. you will go back to the previous look, when you set the first full screen window. 1st is current full screen window 2nd is it proviously geometry 3rd is its type last one is the provious focused window , before that one is set to be full screen.

focused_window: Option<Window>

The vectors of the focus window, when it is empty then there is no focused window, which only can contain 1 object at maximum.

float_windows: Vec<(Window, Geometry)>

all float windows that are under controlled

tile_windows: Vec<Window>

all tiling windows that are under controlled

screen: Screen

we need to know the current size of the screen.

original_windows: Vec<(Window, Geometry)>

original windows with the geometry

minimised_windows: Vec<(Window, Geometry, FloatOrTile)>

minimised windows with the geometry

gap: GapSize

gap