Module cplwm_assignment::g_multiple_workspaces

Optional: Multiple Workspaces

Extend your window manager with support for multiple workspaces. The concept of workspaces is described in the first section of the assignment. See the documentation of the MultiWorkspaceSupport trait for the precise requirements.

Unlike the previous assignments, you are not allowed to make a copy of your previous window manager. You have to define a wrapper implementing the MultiWorkspaceSupport trait. This wrapper can take any existing window manager and uses it to create the different workspaces. This wrapper must also implement all the traits you have implemented in the other assignments, you can forward them to the window manager of the current workspace.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: NO

COMMENTS:

...

Structs

FloatscreenWM
Stuct of Tile screen windows manager

Enums

FloatscreenWMError
The error for tile screen

Type Definitions

WMName
FloatscreenWM