Module cplwm_assignment::h_different_tiling_layout

Optional: Different Tiling Layout

Come up with a different tiling layout algorithm than the one you have already implemented. If you are uninspired, feel free to look for one on the internet, but don't forget to mention where you found it. The layout algorithm may not be trivial, e.g., not just adding tiles by splitting the screen horizontally, and must be at least as complex as, but different enough from the original layout algorithm you already had to implement.

Make a copy of your tiling window manager that implements the tiling layout algorithm. This window manager has to implement the WindowManager trait, but not necessarily the TilingSupport trait, as not every layout has a master tile. Feel free to add additional methods to your window manager that can be used to manipulate its layout. You are not required to let this window manager implement all the previous traits.

Status

TODO: Replace the question mark below with YES, NO, or PARTIAL to indicate the status of this assignment. If you want to tell something about this assignment to the grader, e.g., you have a bug you can't fix, or you want to explain your approach, write it down after the comments section.

COMPLETED: YES

COMMENTS:

The new tiling screen has a master tile and a submaster tile maste till takes the left half of the scrren submaster till takes the top half of rest sub screens, the detail of algorithem will be introduced in the method .

TODO: If you did not come up yourself with this layout, mention its source below.

Structs

TilescreenWM
Stuct of Tile screen windows manager

Enums

TilescreenWMError
The error for tile screen

Type Definitions

WMName
FloatscreenWM