Struct cplwm_assignment::h_different_tiling_layout::TilescreenWM

pub struct TilescreenWM {
    pub windows: Vec<Window>,
    pub focused_window: Vec<Window>,
    pub screen: Screen,
}

Stuct of Tile screen windows manager

Fields
windows: Vec<Window>

The vectors of the windows, the first one is on the master tile

focused_window: Vec<Window>

The vectors of the focus window, when it is empty then there is no focused window, which only can contain 1 object at maximum.

screen: Screen

we need to know the current size of the screen.